Added action to create new business objects. Added window layout to manage mid-span access and splicing

+++ server/branches/version-2.1-spring/util/visual/src/main/java/org/neotropic/util/visual/mxgraph/MxGraphCanvas.java
@@ -203,8 +203,8 @@ public class MxGraphCanvas<N, E> {
             
             newNode.setUuid(nodeId);
             newNode.setLabel(node.toString());
-            newNode.setWidth(Constants.DEFAULT_ICON_WIDTH);
-            newNode.setHeight(Constants.DEFAULT_ICON_HEIGHT);
+            newNode.setWidth((int) Constants.DEFAULT_ICON_WIDTH);
+            newNode.setHeight((int) Constants.DEFAULT_ICON_HEIGHT);
             newNode.setX((xCoordinate)); //The position is scaled
             newNode.setY((yCoordinate));
             newNode.setShape(MxConstants.SHAPE_IMAGE);
